api: reorganize server handlers.
